Center gene data (modified)

Performs per-row centering on a numeric matrix

centerGeneData_new(x, centerGroups = NULL, na.rm = TRUE,
  controlSamples = NULL, useMedian = TRUE, rmOutliers = FALSE,
  madFactor = 5, rowStatsFunc = NULL, returnGroupedValues = FALSE,
  returnGroups = FALSE, verbose = FALSE, ...)

Arguments

x
numeric matrix of input data. See assumptions, that data is assumed to be log2-transformed, or otherwise appropriately transformed.

centerGroups
character vector of group names, or NULL if there are no groups.

na.rm
logical indicating whether NA values should be ignored for summary statistics. This argument is passed to the corresponding row stats function.

controlSamples
character vector of values in colnames(x) which defines the columns to use when calculating group summary values.

useMedian
logical indicating whether to use group median values when calculating summary statistics (TRUE), or group means (FALSE). In either case, when rowStatsFunc is provided, it is used instead.

rmOutliers
logical indicating whether to perform outlier detection and removal prior to row group stats. This argument is passed to jamba::rowGroupMeans(). Note that outliers are only removed during the row group summary step, and not in the centered data.

madFactor
numeric value passed to jamba::rowGroupMeans(), indicating the MAD factor threshold to use when rmOutliers=TRUE. The MAD of each row group is computed, the overall group median MAD is used to define 1x MAD factor, and any MAD more than madFactor times the group median MAD is considered an outlier and is removed. The remaining data is used to compute row group values.

rowStatsFunc
optional function used to calculate row group summary values. This function should take a numeric matrix as input, and return a one-column numeric matrix as output, or a numeric vector with length nrow(x). The function should also accept na.rm as an argument.

returnGroupedValues
logical indicating whether to include the numeric matrix of row group values used during centering, returned in the attributes with name "x_group".

returnGroups
logical indicating whether to return the centering summary data.frame in attributes with name "center_df".

verbose
logical indicating whether to print verbose output.

...
additional arguments are passed to jamba::rowGroupMeans().

Details

This function centers data by subtracting the median or mean for each row.

Optionally columns can be grouped using the argument centerGroups, which will center data within each group independently.

Data can be centered relative to specific control columns using the argument controlSamples. When used with centerGroups, each group of columns defined by centerGroups that does not contain a corresponding value in controlSamples will be centered using the entire group.

Confirm the centerGroups and controlSamples are correct using the attribute "center_df" of the results, see examples below.

Note: This function assumes input data is log2-transformed, or appropriately transformed to fit the assumption of normality. This assumption is necessary for two reasons:

The group value (mean or median) is correct only when data is transformed so the mean or median is not affected by skewed data. Alternatively, rowStatsFunc can be used to specify a custom group summary function.

The centering subtracts the group value from each column value.

See also

Examples

x <- matrix(1:100, ncol=10);
colnames(x) <- letters[1:10];
# basic centering
centerGeneData_new(x);

#> Error in rowGroupMeans(x[, controls_v, drop = FALSE], na.rm = na.rm, groups = centerGroups[controls_v],     useMedian = useMedian, rmOutliers = rmOutliers, madFactor = madFactor,     rowStatsFunc = rowStatsFunc, verbose = verbose, ...): could not find function "rowGroupMeans"

# grouped centering
centerGeneData_new(x,
   centerGroups=rep(c("A","B"), c(5,5)));

#> Error in rowGroupMeans(x[, controls_v, drop = FALSE], na.rm = na.rm, groups = centerGroups[controls_v],     useMedian = useMedian, rmOutliers = rmOutliers, madFactor = madFactor,     rowStatsFunc = rowStatsFunc, verbose = verbose, ...): could not find function "rowGroupMeans"

# centering versus specific control columns
centerGeneData_new(x,
   controlSamples=letters[c(1:3)]);

#> Error in rowGroupMeans(x[, controls_v, drop = FALSE], na.rm = na.rm, groups = centerGroups[controls_v],     useMedian = useMedian, rmOutliers = rmOutliers, madFactor = madFactor,     rowStatsFunc = rowStatsFunc, verbose = verbose, ...): could not find function "rowGroupMeans"

# grouped centering versus specific control columns
centerGeneData_new(x,
   centerGroups=rep(c("A","B"), c(5,5)),
   controlSamples=letters[c(1:3, 6:8)]);

#> Error in rowGroupMeans(x[, controls_v, drop = FALSE], na.rm = na.rm, groups = centerGroups[controls_v],     useMedian = useMedian, rmOutliers = rmOutliers, madFactor = madFactor,     rowStatsFunc = rowStatsFunc, verbose = verbose, ...): could not find function "rowGroupMeans"

# confirm the centerGroups and controlSamples
x_ctr <- centerGeneData_new(x,
   centerGroups=rep(c("A","B"), c(5,5)),
   controlSamples=letters[c(1:3, 6:8)],
   returnGroups=TRUE);

#> Error in rowGroupMeans(x[, controls_v, drop = FALSE], na.rm = na.rm, groups = centerGroups[controls_v],     useMedian = useMedian, rmOutliers = rmOutliers, madFactor = madFactor,     rowStatsFunc = rowStatsFunc, verbose = verbose, ...): could not find function "rowGroupMeans"

attr(x_ctr, "center_df");

#> Error in eval(expr, envir, enclos): object 'x_ctr' not found