Normalize data using MA-plot logic

Normalize data using MA-plot logic

jammanorm(
  x,
  controlGenes = NULL,
  minimum_mean = 0,
  controlSamples = NULL,
  centerGroups = NULL,
  useMedian = FALSE,
  useMean = NULL,
  noise_floor = -Inf,
  noise_floor_value = noise_floor,
  verbose = FALSE,
  ...
)

Arguments

x
numeric matrix with expression data suitable for use by jammaplot(). Gene expression data is typically transformed using log2(1+x) to represent reasonably normal distribution.

minimum_mean
numeric value used to filter controlGenes, a control gene must have at least this level of expression to be included in the normalization, where the expression is determined by the mean or median value analogous to the x-axis value in MA-plots.

controlSamples
character vector of colnames(x) passed to centerGeneData() which defines the control samples during the data centering step. By default, and the most common practice, MA-plots are calculated across all samples, which effectively uses all colnames(x) as controlSamples. However, it is quite useful sometimes to provide a subset of samples especially if there are known quality samples, to which new samples of unknown quality are being compared.

centerGroups
character vector of groups passed to jamma::centerGeneData() which determines how data is centered. Each group is centered independently, to enable visual comparisons within each relevant centering group. It is useful to center within batches or within subsets of samples that are not intended to be compared to one another. Another useful alternative is to center by each sample group in order to view the variability among group replicates, which should be much lower than variability across sample groups. See centerGeneData() for more specific examples.

useMedian
logical indicates whether to center data using the median value, where useMedian=FALSE by default. For consistency, this argument is preferred to useMean which is deprecated and will be removed in future. The median is preferred in cases where outliers should not influence the centering. The mean is preferred in cases where the data should visualize data consistent with downstream parametric statistical analysis. When a particular sample is a technical outlier, one option is to define controlSamples to exclude the outlier sample(s), so the data centering will be applied using the non-outlier samples as reference.

useMean
logical (deprecated), use useMean. This argument indicates whether to center data using the mean value. When useMean=NULL the argument useMedian is preferred. For backward compatibility, when useMean is not NULL, then useMedian is defined by useMedian <- !useMean.

noise_floor, noise_floor_value
numeric values passed to jammacalc(). The noise_floor is the value below which a floor is applied. The floor sets all values below this floor to noise_floor_value. For example, one could apply noise_floor=0 and noise_floor_value=NA which would change any value below 0 to NA.

verbose
logical indicating whether to print verbose output.

...
additional parameters sent to downstream functions, jamba::plotSmoothScatter, centerGeneData.

Value

numeric matrix whose columns have been normalized using the y-axis mean offset following MA-plot calculations. The output contains attributes: "nf" numeric vector of normalization factors; "hk" list of controlGenes used for each sample; "hk_count" the number of controlGenes used for each sample.

Details

This function normalizes data using an approach analogous to data viewed in MA-plots. Normalization is applied by shifting data along the y-axis so the mean (or median) expression among control genes is zero, shown on an MA-plot as y=0.

It effectively reinforces the assumption that the mean log fold change for control genes is expected to be zero. When useMedian=TRUE it reinforces the assumption that the log fold change for the majority of control genes should be zero.

It is useful to run jammaplot() data after jammanorm() to visualize the effect of this normalization. If data is not centered at y=0, the parameters should be adjusted.

About the normalization

This normalization is a "linear normalization" in that it uniformly shifts data up or down relative to other samples, without modifying the relative distribution of signal. It is very similar to housekeeper normalization, geometric mean normalization, and global signal scaling, which are all also "linear normalization" methods. An example of non-linear normalization is quantile or VSN normalization.

The control genes can be defined upfront with controlGenes, which can be housekeeper genes, or a custom subset of genes. The controlGenes are filtered to require the mean or median expression at or above minimum_mean in order to be used during normalization.

The minimum_mean threshold is useful and important to match the variability seen in the MA-plots. For example data below a certain x-axis value may have very high variability, and should usually not be used for normalization.

When the MA-plot after normalization does not show signal centered at y=0, the most common and effective adjustment is to apply minimum_mean to require controGenes to have expression at or above this threshold. The next most effective option is useMedian=TRUE which will center the majority of genes at y=0 instead of the overall mean at y=0.

Comparison to geometric mean normalization

The end result is very similar to other housekeeper normalization methods which typically define normalization factors by calculating the geometric mean of log-transformed housekeeper gene abundances. Such approaches usually work in part because higher expressed housekeeper genes usually have lower variability, which keeps the influence on geometric mean reasonably consistent across a broad range of expression. That said, genes with higher expression have more influence on the geometric mean than genes with much lower expression.

However, the strategy with jammanorm() is to assert that the mean difference from average expression for the controlGenes should be equal to zero. The effect is applied evenly across control genes by evaluating the mean or median difference from y=0 for each sample.

Assumptions

In order for this approach to be valid:

Data on the MA-plots should be horizontal for all samples, particularly for controlGenes. When data is not horizontal across samples, data should instead be normalized using another approach, typically something like quantile normalization which is intended to impose a consistent signal distribution across all samples.

The majority of controlGenes should not be changing, therefore the controlGenes should not have substantial variation along the y-axis within each sample.

For data with a small number of genes, normalization should use a set of pre-defined controlGenes that were verified to have no change across samples.

Noise threshold

Note that some platform technologies generate a noise threshold, below which data may be skewed up or down relative to other samples. It is recommended to ignore this type of skew below the threshold when determining whether data is horizontal on MA-plots.

For example Nanostring data includes a series of positive and negative control probes, and a suitable noise threshold is either the midpoint between the lowest positive and highest negative probe, or the lowest positive probe. When this noise threshold is applied, data above the noise threshold is typically horizontal, although data below the threshold may be skewed up or down depending upon the effective input RNA concentration.

See also