Find colname by string or pattern

Find colname by string or pattern, with option to require non-NA values.

find_colname(
  pattern,
  x,
  max = 1,
  index = FALSE,
  require_non_na = TRUE,
  col_types = NULL,
  exclude_pattern = NULL,
  verbose = FALSE,
  ...
)

Arguments

pattern

character vector of text strings and/or regular expression patterns.

x

data.frame or other object that contains colnames(x).

max

integer maximum number of entries to return.

index

logical indicating whether to return the column index, that is the column number.

require_non_na

logical indicating whether to require at least one non-NA value in the matching colname. When require_non_na=TRUE and all values in a column are NA, that colname is not returned by this function.

exclude_pattern

character vector of colnames or patterns to exclude from returned results.

verbose

logical indicating whether to print verbose output.

...

additional arguments are ignored.

Details

This function is a simple utility function intended to help find the most appropriate matching colname given one or more character strings or patterns.

It returns the first best matching result, but can return multiple results in order of preference if max=Inf.

The order of matching:

Match the exact colname.

Match case-insensitive colname.

Match the beginning of each colname.

Match the end of each colname.

Match anywhere in each colname.

The goal is to use something like c("p.value", "pvalue", "pval") and be able to find colnames with these variations:

P.Value

P.Value Group-Control

Group-Control P.Value

pvalue

Even if the data contains c("P.Value", "adj.P.Val") as returned by limma::topTable() for example, the pattern c("p.val") will preferentially match "P.Value" and not "adj.P.Val".

Examples

x <- data.frame(
   `Gene`=paste0("gene", LETTERS[1:25]),
   `log2fold Group-Control`=rnorm(25)*2,
   `P.Value Group-Control`=10^-rnorm(25)^2,
   check.names=FALSE);
x[["fold Group-Control"]] <- log2fold_to_fold(x[["log2fold Group-Control"]]);
x[["adj.P.Val Group-Control"]] <- x[["P.Value Group-Control"]];

print(head(x));
#>    Gene log2fold Group-Control P.Value Group-Control fold Group-Control
#> 1 geneA              1.1781147            0.59374671           2.262809
#> 2 geneB             -3.6698269            0.84401845         -12.727057
#> 3 geneC              0.0842527            0.25708661           1.060138
#> 4 geneD             -0.4569238            0.38992803          -1.372612
#> 5 geneE             -1.0217770            0.04991258          -2.030418
#> 6 geneF             -1.5149908            0.47738895          -2.857970
#>   adj.P.Val Group-Control
#> 1              0.59374671
#> 2              0.84401845
#> 3              0.25708661
#> 4              0.38992803
#> 5              0.04991258
#> 6              0.47738895
find_colname(c("p.val", "pval"), x);
#> [1] "P.Value Group-Control"
find_colname(c("fold", "fc", "ratio"), x);
#> [1] "fold Group-Control"
find_colname(c("logfold", "log2fold", "lfc", "log2ratio", "logratio"), x);
#> [1] "log2fold Group-Control"

## use exclude_pattern
## if the input data has no "P.Value" but has "adj.P.Val"
y <- x[,c(1,2,4,5)];
print(head(y));
#>    Gene log2fold Group-Control fold Group-Control adj.P.Val Group-Control
#> 1 geneA              1.1781147           2.262809              0.59374671
#> 2 geneB             -3.6698269         -12.727057              0.84401845
#> 3 geneC              0.0842527           1.060138              0.25708661
#> 4 geneD             -0.4569238          -1.372612              0.38992803
#> 5 geneE             -1.0217770          -2.030418              0.04991258
#> 6 geneF             -1.5149908          -2.857970              0.47738895
find_colname(c("p.val"), y, exclude_pattern=c("adj"))
#> NULL